[zero] Move extendTheme to already existing @mui/zero-runtime/utils
path. Doesn't make sense to have so many export subpaths since these
utils are mostly there to be consumed by bundler packages.
