+++ packages/zero-runtime/.gitignore
@@ -1,4 +1,3 @@
 /processors/
 /utils/
-/extendTheme/
 LICENSE
