+++ packages/zero-runtime/package.json
@@ -21,7 +21,7 @@
     "url": "https://opencollective.com/mui-org"
   },
   "scripts": {
-    "clean": "rimraf build extendTheme types processors utils",
+    "clean": "rimraf build types processors utils",
     "watch": "tsup --watch --clean false",
     "copy-license": "node ../../scripts/zero-runtime-license.mjs",
     "build": "tsup",
@@ -82,7 +82,6 @@
   "files": [
     "build",
     "exports",
-    "extendTheme",
     "processors",
     "theme",
     "utils",
@@ -102,15 +101,6 @@
       "import": "./theme/index.mjs",
       "require": "./theme/index.js",
       "default": "./theme/index.js"
-    },
-    "./extendTheme": {
-      "types": "./extendTheme/index.d.ts",
-      "import": {
-        "default": "./extendTheme/index.mjs",
-        "types": "./extendTheme/index.d.mts"
-      },
-      "require": "./extendTheme/index.js",
-      "default": "./extendTheme/index.js"
     },
     "./styles.css": {
       "default": "./styles.css"
@@ -144,7 +134,6 @@
       "build": {
         "outputs": [
           "{projectRoot}/build",
-          "{projectRoot}/extendTheme",
           "{projectRoot}/processors",
           "{projectRoot}/utils"
         ]
