+++ packages/zero-runtime/src/theme.ts
@@ -1,1 +1,2 @@
 export interface ThemeArgs {}
+export { ExtendTheme } from './utils/extendTheme';
