+++ packages/zero-runtime/src/utils/extendTheme.ts
@@ -1,7 +1,7 @@
 import deepMerge from 'lodash/merge';
 import { prepareCssVars } from '@mui/system/cssVars';
 import type { SxConfig } from '@mui/system/styleFunctionSx';
-import type { CSSObject } from './base';
+import type { CSSObject } from '../base';
 
 export interface ThemeInput<ColorScheme extends string = string> {
   /**
