+++ packages/zero-runtime/src/utils/index.ts
@@ -1,3 +1,4 @@
 export type { PluginCustomOptions } from './cssFnValueToVariable';
 export * from './preprocessor';
 export * from './generateCss';
+export * from './extendTheme';
