+++ packages/zero-unplugin/src/index.ts
@@ -18,8 +18,9 @@ import {
   preprocessor as basePreprocessor,
   generateTokenCss,
   generateThemeTokens,
+  extendTheme,
+  type Theme as BaseTheme,
 } from '@mui/zero-runtime/utils';
-import { extendTheme, type Theme as BaseTheme } from '@mui/zero-runtime/extendTheme';
 
 type NextMeta = {
   type: 'next';
